requirements.txt
anyio==3.6.2
astroid==2.12.13
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
# iqoptionapi==7.1.1
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.6.0
py-iqoption-api==1.1.4
pycodestyle==2.10.0
pydantic==1.10.2
pylint==2.15.8
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket-client==0.56.0
websockets==10.4
wrapt==1.14.1
